remove extraneous - flag in test.
Maybe I meant to type '+' but '-' is not defined for this package.  The
code doesn't catch it and so ends up passing the '-' through to the fmt
package.  There, the behavior has changed at tip vs. 1.6 and '-' now
causes undesired results that fail the example.  ...  Anyway, '-' was
incorrect in this test.